+++ src/features/app/App.jsx
@@ -8,7 +8,6 @@ import UserGoals from '../usergoals/UserGoals';
 import { setToken } from '../user/userSlice';
 import Header from './Header';
 import { getToken } from '../../app/utils';
-import './background-style.scss';
 import Team from '../team/Team';
 import Signin from '../user/Signin';
 import Signup from '../user/Signup';
@@ -20,22 +19,6 @@ function FallBack() {
   return (
     <div>
       <h1>404 Not Found</h1>
-    </div>
-  );
-}
-
-function Circles() {
-  return (
-    <div className="circles">
-      <div className="circle a" />
-      <div className="circle b" />
-      <div className="circle c" />
-      <div className="circle d" />
-      <div className="circle e" />
-      <div className="circle f" />
-      <div className="circle g" />
-      <div className="circle h" />
-      <div className="circle_cover" />
     </div>
   );
 }
@@ -83,7 +66,7 @@ function AppWrapper() {
   return (
     <BrowserRouter>
       <div>
-        <Circles />
+        {/* <Circles /> */}
         <App />
       </div>
     </BrowserRouter>
