+++ src/features/usergoals/SustainabilityGoalInput.jsx
@@ -9,7 +9,7 @@ function SustyGoalInput() {
     dispatch(getGoals());
   }, []);
   return (
-    <div className="container">
+    <div className="container background">
       <h3>What is your sustainability goal?</h3>
       <input value={goal} type="text" onChange={(e) => { setGoalState(e.target.value); }} />
       <button type="button" className="button" onClick={() => { dispatch(setGoal({ description: goal })); setGoalState(''); }}>
