+++ src/features/usergoals/UserGoalSelection.jsx
@@ -7,9 +7,9 @@ function UserGoalSelection(props) {
   const dispatch = useDispatch();
 
   return (
-    <div className="container center">
-      <h2>Goal {props.index + 1}</h2>
-      <h3>{props.goal.description}</h3>
+    <div className="container center background">
+      <h2 className="post-title">Goal {props.index + 1}</h2>
+      <h5 className="goaltext">{props.goal.description}</h5>
       <h3>Did you complete your goal today?</h3>
       <button type="button" className="button" onClick={() => { dispatch(completeGoal(props.goal.id)); }}>
         Yes!
@@ -17,10 +17,10 @@ function UserGoalSelection(props) {
       <button type="button" className="button" onClick={() => { dispatch(deleteGoal(props.goal.id)); }}>
         No!
       </button>
+      <Streak />
       <button type="button" className="button" onClick={() => { dispatch(deleteGoal(props.goal.id)); }}>
         Delete Goal
       </button>
-      <Streak />
     </div>
   );
 }
