+++ src/figmastyle.scss
@@ -1,0 +1,13 @@
+#header {
+    width: 100%;
+    height: $header-height;
+    background-color: $secondary-color-semitranslucent;
+    //display: flex;
+    align-items: center;
+    backdrop-filter: blur(5px);
+    border-bottom: 0.75px solid $border-color;
+    font-family: 'Oswald', sans-serif;
+    position: fixed;
+    display: flex;
+    top: 0;
+  }
