+++ src/style.scss
@@ -1,7 +1,8 @@
 
-$background-color: #131b20;
+$background-color: #D9D9D9;
 $background-color-translucent: rgba(19, 27, 32, 0.369);
 $primary-color: #222c32;
+$dartmouth-green: #00693e;
 $primary-color-translucent: rgba(34, 44, 50, 0.5);
 $primary-color-translucent-unselected: rgba(255, 13, 0, 0.31);
 $secondary-color: #323d49;
@@ -12,24 +13,16 @@ $header-height: 100px;
 $nav-width: 200px;
 $opaqueness: #ffffff00;
 
-#header {
-  width: 100svw;
-  height: $header-height;
-  background-color: $secondary-color-semitranslucent;
-  //display: flex;
-  align-items: center;
-  backdrop-filter: blur(5px);
-  border-bottom: 0.75px solid $border-color;
-  font-family: 'Oswald', sans-serif;
-  position: fixed;
-  display: flex;
-  top: 0;
+/* Rectangle 27 */
+
+.background{
+  background-color: #FFFFFF;
 }
 
 #nav {
-  height: 100%;
+  height: 100svh;
   width: $nav-width;
-  background-color: $secondary-color-semitranslucent;
+  background-color: $dartmouth-green;
   backdrop-filter: blur(5px);
   display: flex;
   flex-direction: column;
@@ -38,11 +31,28 @@ $opaqueness: #ffffff00;
   position: fixed;
 }
 
+#header {
+  width: 100%;
+  height: $header-height;
+  background-color: $background-color;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  //display: flex;
+  align-items: center;
+  // backdrop-filter: blur(5px);
+  border-bottom: 0.75px solid $border-color;
+  font-family: 'Oswald', sans-serif;
+  position: fixed;
+  display: flex;
+  top: 0;
+}
+
+
 #app {
   display: flex;
   flex-direction: row;
-  margin-top: $header-height;
+  margin-top: 120px;
   padding-left: $nav-width;
+  
 }
 
 .page {
@@ -101,7 +111,7 @@ textarea, input, p, h1, h2, h3, h4, h5{
   box-shadow: none;
 
   resize: none;
-  color: white;
+  color:black;
   margin: 10px;
 
   font-family: 'Oswald', sans-serif;
@@ -128,7 +138,7 @@ h1 {
 
 h2{
   font-size: 2.5em;
-  margin: 0px;
+  margin-left: 10px;
 }
 
 h3 {
@@ -137,6 +147,11 @@ h3 {
 
 h4 {
   font-size: 1em;
+}
+
+.goaltext{
+  font-size: 2em;
+  color: $dartmouth-green
 }
 
 .allign-v{
@@ -153,10 +168,12 @@ h4 {
 
 .post-title {
   font-size: 1.5em;
-  width: 100%;
-  height: 50px;
-  background-color: $secondary-color-translucent;
-  margin: 0px;
+  // width: 100%;
+  height: 30px;
+  background-color: rgba(5, 133, 24, 0.25);;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  border-radius: 8px;
+  margin: 10px;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -174,13 +191,13 @@ h4 {
 }
 
 .container{
-  background-color: $primary-color-translucent;
+  // background-color: $primary-color-translucent;
   border-radius: 8px;
   overflow: hidden;
-  border: 0.75px solid $border-color;
-  margin: 40px auto;
+  // border: 0.75px solid $border-color;
+  margin: 10px;
   transition: transform 0.5s, box-shadow 0.5s;
-  backdrop-filter: blur(5px);
+  // backdrop-filter: blur(5px);
   width: calc(100% - 80px);
 }
 
@@ -191,7 +208,8 @@ h4 {
   display: flex;
   // flex-direction: column;
   justify-content: center;
-  padding:40px;
+  // padding:40px;
+  border:5px
 }
 
 .post{
